requirements.txt
asgiref==3.11.1
google-auth==2.49.1
async-timeout==5.0.1
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
boto3==1.42.56
botocore==1.42.56
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
constantly==23.10.4
cryptography==46.0.5
daphne==4.2.1
Django==5.2.11
django-cors-headers==4.9.0
django-extensions==4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
jmespath==1.1.0
msgpack==1.1.2
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.11.0
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.2.0
s3transfer==0.16.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
tomli==2.4.0
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
urllib3==2.6.3
zope.interface==8.2
requests
razorpay==1.4.2
gunicorn==23.0.0
whitenoise==6.9.0
django_redis
jcodemunch-mcp
